* libF77/getenv_.c (G77_getenv_0): Avoid signed/unsigned warning. 	* libF77/system_.c (G77_system_0): Likewise. 	* libI77/open.c (f_open): Likewise. 	* libI77/rdfmt.c (rd_Z): Likewise. 	* libI77/uio.c (do_us, do_ud): Likewise.
git-svn-id: svn+ssh://gcc.gnu.org/svn/gcc/trunk@54166 138bc75d-0d04-0410-961f-82ee72b054a4

+++ libf2c/libI77/open.c
@@ -199,7 +199,7 @@ f_open (olist * a)
       if (!env)
 	env = "/tmp";
       len = strlen (env);
-      if (len > 256 - sizeof "/tmp.FXXXXXX")
+      if (len > 256 - (int) sizeof ("/tmp.FXXXXXX"))
 	err (a->oerr, 132, "open");
       strcpy (buf, env);
       strcat (buf, "/tmp.FXXXXXX");
